+++ photutils/segmentation/properties.py
@@ -1,12 +1,16 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
+import warnings
 
 import numpy as np
 import astropy.units as u
-from astropy.table import Table
-from astropy.utils import lazyproperty
+from astropy.coordinates import SkyCoord
+from astropy.table import QTable
+from astropy.utils import deprecated, lazyproperty
+from astropy.utils.exceptions import AstropyDeprecationWarning
 from astropy.wcs.utils import pixel_to_skycoord
+
 
 from .core import SegmentationImage
 from ..utils.convolution import filter_data
@@ -76,10 +80,8 @@ class SourceProperties(object):
         background-subtracted.
 
     wcs : `~astropy.wcs.WCS`
-        The WCS transformation to use.  If `None`, then
-        `~photutils.SourceProperties.icrs_centroid`,
-        `~photutils.SourceProperties.ra_icrs_centroid`, and
-        `~photutils.SourceProperties.dec_icrs_centroid` will be `None`.
+        The WCS transformation to use.  If `None`, then any sky-based
+        properties will be set to `None`.
 
     Notes
     -----
@@ -213,10 +215,10 @@ class SourceProperties(object):
 
     def to_table(self, columns=None, exclude_columns=None):
         """
-        Create a `~astropy.table.Table` of properties.
+        Create a `~astropy.table.QTable` of properties.
 
         If ``columns`` or ``exclude_columns`` are not input, then the
-        `~astropy.table.Table` will include all scalar-valued
+        `~astropy.table.QTable` will include all scalar-valued
         properties.  Multi-dimensional properties, e.g.
         `~photutils.SourceProperties.data_cutout`, can be included in
         the ``columns`` input.
@@ -225,17 +227,17 @@ class SourceProperties(object):
         ----------
         columns : str or list of str, optional
             Names of columns, in order, to include in the output
-            `~astropy.table.Table`.  The allowed column names are any of
-            the attributes of `SourceProperties`.
+            `~astropy.table.QTable`.  The allowed column names are any
+            of the attributes of `SourceProperties`.
 
         exclude_columns : str or list of str, optional
             Names of columns to exclude from the default properties list
-            in the output `~astropy.table.Table`.  The default
+            in the output `~astropy.table.QTable`.  The default
             properties are those with scalar values.
 
         Returns
         -------
-        table : `~astropy.table.Table`
+        table : `~astropy.table.QTable`
             A single-row table of properties of the source.
         """
 
@@ -418,21 +420,47 @@ class SourceProperties(object):
         return self.centroid[0]
 
     @lazyproperty
-    def icrs_centroid(self):
+    def sky_centroid(self):
         """
-        The International Celestial Reference System (ICRS) coordinates
-        of the centroid within the source segment, returned as a
-        `~astropy.coordinates.SkyCoord` object.
+        The sky coordinates of the centroid within the source segment,
+        returned as a `~astropy.coordinates.SkyCoord` object.
+
+        The output coordinate frame is the same as the input WCS.
         """
 
         if self._wcs is not None:
             return pixel_to_skycoord(self.xcentroid.value,
                                      self.ycentroid.value,
-                                     self._wcs, origin=1).icrs
+                                     self._wcs, origin=0)
         else:
             return None
 
     @lazyproperty
+    def sky_centroid_icrs(self):
+        """
+        The sky coordinates, in the International Celestial Reference
+        System (ICRS) frame, of the centroid within the source segment,
+        returned as a `~astropy.coordinates.SkyCoord` object.
+        """
+
+        if self._wcs is not None:
+            return self.sky_centroid.icrs
+        else:
+            return None
+
+    @lazyproperty
+    @deprecated(0.4, alternative='sky_centroid_icrs')
+    def icrs_centroid(self):
+        """
+        The sky coordinates, in the International Celestial Reference
+        System (ICRS) frame, of the centroid within the source segment,
+        returned as a `~astropy.coordinates.SkyCoord` object.
+        """
+
+        return self.sky_centroid_icrs
+
+    @lazyproperty
+    @deprecated(0.4, alternative='sky_centroid_icrs.ra')
     def ra_icrs_centroid(self):
         """
         The ICRS Right Ascension coordinate (in degrees) of the centroid
@@ -440,11 +468,12 @@ class SourceProperties(object):
         """
 
         if self._wcs is not None:
-            return self.icrs_centroid.ra.degree * u.deg
+            return self.sky_centroid_icrs.ra.degree * u.deg
         else:
             return None
 
     @lazyproperty
+    @deprecated(0.4, alternative='sky_centroid_icrs.dec')
     def dec_icrs_centroid(self):
         """
         The ICRS Declination coordinate (in degrees) of the centroid
@@ -452,7 +481,7 @@ class SourceProperties(object):
         """
 
         if self._wcs is not None:
-            return self.icrs_centroid.dec.degree * u.deg
+            return self.sky_centroid_icrs.dec.degree * u.deg
         else:
             return None
 
@@ -502,6 +531,78 @@ class SourceProperties(object):
         """
 
         return self.bbox[2]
+
+    @lazyproperty
+    def sky_bbox_ll(self):
+        """
+        The sky coordinates of the lower-left vertex of the minimal
+        bounding box of the source segment, returned as a
+        `~astropy.coordinates.SkyCoord` object.
+
+        The bounding box encloses all of the source segment pixels in
+        their entirety, thus the vertices are at the pixel *corners*.
+        """
+
+        if self._wcs is not None:
+            return pixel_to_skycoord(self.xmin.value - 0.5,
+                                     self.ymin.value - 0.5,
+                                     self._wcs, origin=0)
+        else:
+            return None
+
+    @lazyproperty
+    def sky_bbox_ul(self):
+        """
+        The sky coordinates of the upper-left vertex of the minimal
+        bounding box of the source segment, returned as a
+        `~astropy.coordinates.SkyCoord` object.
+
+        The bounding box encloses all of the source segment pixels in
+        their entirety, thus the vertices are at the pixel *corners*.
+        """
+
+        if self._wcs is not None:
+            return pixel_to_skycoord(self.xmin.value - 0.5,
+                                     self.ymax.value + 0.5,
+                                     self._wcs, origin=0)
+        else:
+            return None
+
+    @lazyproperty
+    def sky_bbox_lr(self):
+        """
+        The sky coordinates of the lower-right vertex of the minimal
+        bounding box of the source segment, returned as a
+        `~astropy.coordinates.SkyCoord` object.
+
+        The bounding box encloses all of the source segment pixels in
+        their entirety, thus the vertices are at the pixel *corners*.
+        """
+
+        if self._wcs is not None:
+            return pixel_to_skycoord(self.xmax.value + 0.5,
+                                     self.ymin.value - 0.5,
+                                     self._wcs, origin=0)
+        else:
+            return None
+
+    @lazyproperty
+    def sky_bbox_ur(self):
+        """
+        The sky coordinates of the upper-right vertex of the minimal
+        bounding box of the source segment, returned as a
+        `~astropy.coordinates.SkyCoord` object.
+
+        The bounding box encloses all of the source segment pixels in
+        their entirety, thus the vertices are at the pixel *corners*.
+        """
+
+        if self._wcs is not None:
+            return pixel_to_skycoord(self.xmax.value + 0.5,
+                                     self.ymax.value + 0.5,
+                                     self._wcs, origin=0)
+        else:
+            return None
 
     @lazyproperty
     def min_value(self):
@@ -1003,10 +1104,8 @@ def source_properties(data, segment_img, error=None, mask=None,
         calculated from the filtered "detection" image.
 
     wcs : `~astropy.wcs.WCS`
-        The WCS transformation to use.  If `None`, then
-        `~photutils.SourceProperties.icrs_centroid`,
-        `~photutils.SourceProperties.ra_icrs_centroid`, and
-        `~photutils.SourceProperties.dec_icrs_centroid` will be `None`.
+        The WCS transformation to use.  If `None`, then any sky-based
+        properties will be set to `None`.
 
     labels : int, array-like (1D, int)
         Subset of segmentation labels for which to calculate the
@@ -1131,12 +1230,12 @@ def source_properties(data, segment_img, error=None, mask=None,
 
 def properties_table(source_props, columns=None, exclude_columns=None):
     """
-    Construct a `~astropy.table.Table` of properties from a list of
+    Construct a `~astropy.table.QTable` of properties from a list of
     `SourceProperties` objects.
 
     If ``columns`` or ``exclude_columns`` are not input, then the
-    `~astropy.table.Table` will include all scalar-valued properties.
-    Multi-dimensional properties, e.g.
+    `~astropy.table.QTable` will include most scalar-valued source
+    properties.  Multi-dimensional properties, e.g.
     `~photutils.SourceProperties.data_cutout`, can be included in the
     ``columns`` input.
 
@@ -1148,17 +1247,17 @@ def properties_table(source_props, columns=None, exclude_columns=None):
 
     columns : str or list of str, optional
         Names of columns, in order, to include in the output
-        `~astropy.table.Table`.  The allowed column names are any of the
-        attributes of `SourceProperties`.
+        `~astropy.table.QTable`.  The allowed column names are any of
+        the attributes of `SourceProperties`.
 
     exclude_columns : str or list of str, optional
         Names of columns to exclude from the default properties list in
-        the output `~astropy.table.Table`.  The default properties are
+        the output `~astropy.table.QTable`.  The default properties are
         those with scalar values.
 
     Returns
     -------
-    table : `~astropy.table.Table`
+    table : `~astropy.table.QTable`
         A table of properties of the segmented sources, one row per
         source.
 
@@ -1183,12 +1282,14 @@ def properties_table(source_props, columns=None, exclude_columns=None):
     >>> props = source_properties(image, segm)
     >>> columns = ['id', 'xcentroid', 'ycentroid', 'source_sum']
     >>> tbl = properties_table(props, columns=columns)
+    >>> tbl['xcentroid'].info.format = '.10f'  # optional format
+    >>> tbl['ycentroid'].info.format = '.10f'  # optional format
     >>> print(tbl)
-     id   xcentroid     ycentroid   source_sum
-             pix           pix
-    --- ------------- ------------- ----------
-      1           0.2           0.8        5.0
-      2 2.09090909091 2.36363636364       55.0
+     id  xcentroid    ycentroid   source_sum
+            pix          pix
+    --- ------------ ------------ ----------
+      1 0.2000000000 0.8000000000        5.0
+      2 2.0909090909 2.3636363636       55.0
     """
 
     if isinstance(source_props, list) and len(source_props) == 0:
@@ -1196,8 +1297,8 @@ def properties_table(source_props, columns=None, exclude_columns=None):
     source_props = np.atleast_1d(source_props)
 
     # all scalar-valued properties
-    columns_all = ['id', 'xcentroid', 'ycentroid', 'ra_icrs_centroid',
-                   'dec_icrs_centroid', 'source_sum',
+    columns_all = ['id', 'xcentroid', 'ycentroid', 'sky_centroid',
+                   'sky_centroid_icrs', 'source_sum',
                    'source_sum_err', 'background_sum', 'background_mean',
                    'background_at_centroid', 'xmin', 'xmax', 'ymin', 'ymax',
                    'min_value', 'max_value', 'minval_xpos', 'minval_ypos',
@@ -1215,39 +1316,100 @@ def properties_table(source_props, columns=None, exclude_columns=None):
     if table_columns is None:
         table_columns = columns_all
 
-    # it's *much* faster to calculate world coordinates using the
-    # complete list of (x, y) instead of from the individual (x, y).
-    # The assumption here is that the wcs is the same for each
+    # It's *much* faster to calculate world coordinates using the
+    # complete list of (x, y) instead of looping through the individual
+    # (x, y).  It's also much faster to recalculate world coordindates
+    # than to create a SkyCoord array from a loop-generated SkyCoord
+    # list.  The assumption here is that the wcs is the same for each
     # element of source_props.
-    if ('ra_icrs_centroid' in table_columns or
-            'dec_icrs_centroid' in table_columns or
-            'icrs_centroid' in table_columns):
-        xcentroid = [props.xcentroid.value for props in source_props]
-        ycentroid = [props.ycentroid.value for props in source_props]
+    sky_colnames = ['sky_centroid', 'sky_centroid_icrs', 'icrs_centroid',
+                    'ra_icrs_centroid', 'dec_icrs_centroid']
+    calc_skycoords = any(sky_colname in table_columns
+                         for sky_colname in sky_colnames)
+
+    if calc_skycoords:
         if source_props[0]._wcs is not None:
-            icrs_centroid = pixel_to_skycoord(
-                xcentroid, ycentroid, source_props[0]._wcs, origin=1).icrs
-            icrs_ra = icrs_centroid.ra.degree * u.deg
-            icrs_dec = icrs_centroid.dec.degree * u.deg
+            xcentroid = [props.xcentroid.value for props in source_props]
+            ycentroid = [props.ycentroid.value for props in source_props]
+
+            sky_centroid = pixel_to_skycoord(
+                xcentroid, ycentroid, source_props[0]._wcs, origin=0)
+            sky_centroid_icrs = sky_centroid.icrs
+            icrs_centroid = sky_centroid_icrs
+            ra_icrs_centroid = sky_centroid_icrs.ra.deg * u.deg
+            dec_icrs_centroid = sky_centroid_icrs.dec.deg * u.deg
         else:
             nprops = len(source_props)
-            icrs_ra = [None] * nprops
-            icrs_dec = [None] * nprops
+            sky_centroid = sky_centroid_icrs = [None] * nprops
             icrs_centroid = [None] * nprops
+            ra_icrs_centroid = dec_icrs_centroid = [None] * nprops
 
-    props_table = Table()
+    bbox_colnames = ['sky_bbox_ll', 'sky_bbox_ul', 'sky_bbox_lr',
+                     'sky_bbox_ur']
+    calc_bboxcoords = any(bbox_colname in table_columns
+                          for bbox_colname in bbox_colnames)
+
+    if calc_bboxcoords:
+        if source_props[0]._wcs is not None:
+            xmin = np.array([props.xmin.value for props in source_props])
+            ymin = np.array([props.ymin.value for props in source_props])
+            xmax = np.array([props.xmax.value for props in source_props])
+            ymax = np.array([props.ymax.value for props in source_props])
+
+            wcs = source_props[0]._wcs
+            sky_bbox_ll = pixel_to_skycoord(xmin - 0.5, ymin - 0.5, wcs,
+                                            origin=0)
+            sky_bbox_ul = pixel_to_skycoord(xmin - 0.5, ymax + 0.5, wcs,
+                                            origin=0)
+            sky_bbox_lr = pixel_to_skycoord(xmax + 0.5, ymin - 0.5, wcs,
+                                            origin=0)
+            sky_bbox_ur = pixel_to_skycoord(xmax + 0.5, ymax + 0.5, wcs,
+                                            origin=0)
+        else:
+            nprops = len(source_props)
+            sky_bbox_ll = sky_bbox_ul = [None] * nprops
+            sky_bbox_lr = sky_bbox_ur = [None] * nprops
+
+    props_table = QTable()
     for column in table_columns:
-        if column == 'ra_icrs_centroid':
-            props_table[column] = icrs_ra
-        elif column == 'dec_icrs_centroid':
-            props_table[column] = icrs_dec
+        if column == 'sky_centroid':
+            props_table[column] = sky_centroid
+        elif column == 'sky_centroid_icrs':
+            props_table[column] = sky_centroid_icrs
         elif column == 'icrs_centroid':
+            warnings.warn('The icrs_centroid property is deprecated and '
+                          'may be removed in a future version.  Use '
+                          'sky_centroid_icrs instead',
+                          AstropyDeprecationWarning)
             props_table[column] = icrs_centroid
+        elif column == 'ra_icrs_centroid':
+            warnings.warn('The ra_icrs_centroid property is deprecated and '
+                          'may be removed in a future version.  Use '
+                          'sky_centroid_icrs.ra instead',
+                          AstropyDeprecationWarning)
+            props_table[column] = ra_icrs_centroid
+        elif column == 'dec_icrs_centroid':
+            warnings.warn('The dec_icrs_centroid property is deprecated and '
+                          'may be removed in a future version.  Use '
+                          'sky_centroid_icrs.dec instead',
+                          AstropyDeprecationWarning)
+            props_table[column] = dec_icrs_centroid
+        elif column == 'sky_bbox_ll':
+            props_table[column] = sky_bbox_ll
+        elif column == 'sky_bbox_ul':
+            props_table[column] = sky_bbox_ul
+        elif column == 'sky_bbox_lr':
+            props_table[column] = sky_bbox_lr
+        elif column == 'sky_bbox_ur':
+            props_table[column] = sky_bbox_ur
         else:
             values = [getattr(props, column) for props in source_props]
             if isinstance(values[0], u.Quantity):
                 # turn list of Quantities into a Quantity array
                 values = u.Quantity(values)
+            if isinstance(values[0], SkyCoord):   # failsafe
+                values = SkyCoord(values)
+
             props_table[column] = values
 
     return props_table